+++ src/UsbSentinel.Service/Program.cs
@@ -14,5 +14,6 @@ builder.Services.AddSingleton<SentinelCoordinator>();
 builder.Services.AddSingleton<PipeServer>();
 builder.Services.AddHostedService<SentinelWorker>();
 builder.Services.AddHostedService<DefenderSignatureWorker>();
+builder.Services.AddHostedService<ProtectionAuditWorker>();
 
 await builder.Build().RunAsync();